+++ packages/common/src/types.ts
@@ -1552,8 +1552,9 @@ export interface IdentityInput {
 }
 
 export enum SchemaVersion {
-    V1 = 0, // Used by version 0 smart contracts.
-    V2 = 1, // Used by version 1 smart contracts.
+    V0 = 0, // Used by version 0 smart contracts.
+    V1 = 1, // Used by version 1 smart contracts.
+    V2 = 2, // Used by version 1 smart contracts.
 }
 
 export type IpArData = {