+++ packages/common/test/deserializeSchema.test.ts
@@ -3,15 +3,15 @@ import fs from 'fs';
 import { deserialModuleFromBuffer } from '../src/deserializeSchema';
 import { SchemaVersion } from '../src/types';
 
-test('Deserialize V1 schema from file (two-step-transfer)', async () => {
+test('Deserialize V0 schema from file (two-step-transfer)', async () => {
     const rawModule = Buffer.from(
         fs.readFileSync('./test/resources/two-step-transfer-schema.bin')
     );
-    const parsedModule = deserialModuleFromBuffer(rawModule, SchemaVersion.V1);
+    const parsedModule = deserialModuleFromBuffer(rawModule, SchemaVersion.V0);
 
     const contractName = 'two-step-transfer';
 
-    if (parsedModule.v != SchemaVersion.V1) {
+    if (parsedModule.v != SchemaVersion.V0) {
         throw new Error('Unexpected schema version');
     }
 
@@ -20,11 +20,11 @@ test('Deserialize V1 schema from file (two-step-transfer)', async () => {
     expect(parsedModule.value[contractName].receive).toBeDefined();
 });
 
-test('Deserialize V2 schema from file (icecream)', async () => {
+test('Deserialize V1 schema from file (icecream)', async () => {
     const rawModule = Buffer.from(
         fs.readFileSync('./test/resources/icecream-schema.bin')
     );
-    const parsedModule = deserialModuleFromBuffer(rawModule, SchemaVersion.V2);
+    const parsedModule = deserialModuleFromBuffer(rawModule, SchemaVersion.V1);
 
     const contractName = 'icecream';
 
@@ -32,15 +32,15 @@ test('Deserialize V2 schema from file (icecream)', async () => {
     expect(parsedModule.value[contractName].receive).toBeDefined();
 });
 
-test('Deserialize V1 versioned schema from file (cis1-wccd)', async () => {
+test('Deserialize V0 versioned schema from file (cis1-wccd)', async () => {
     const rawModule = Buffer.from(
-        fs.readFileSync('./test/resources/cis1-wccd-schema-v1-versioned.bin')
+        fs.readFileSync('./test/resources/cis1-wccd-schema-v0-versioned.bin')
     );
     const parsedModule = deserialModuleFromBuffer(rawModule);
 
     const contractName = 'CIS1-wCCD';
 
-    if (parsedModule.v != SchemaVersion.V1) {
+    if (parsedModule.v != SchemaVersion.V0) {
         throw new Error('Unexpected schema version');
     }
 
@@ -49,13 +49,29 @@ test('Deserialize V1 versioned schema from file (cis1-wccd)', async () => {
     expect(parsedModule.value[contractName].receive).toBeDefined();
 });
 
-test('Deserialize V2 versioned schema from file (cis2-wccd)', async () => {
+test('Deserialize V1 versioned schema from file (cis2-wccd)', async () => {
     const rawModule = Buffer.from(
-        fs.readFileSync('./test/resources/cis2-wccd-schema-v2-versioned.bin')
+        fs.readFileSync('./test/resources/cis2-wccd-schema-v1-versioned.bin')
     );
     const parsedModule = deserialModuleFromBuffer(rawModule);
 
     const contractName = 'CIS2-wCCD';
+
+    if (parsedModule.v != SchemaVersion.V1) {
+        throw new Error('Unexpected schema version');
+    }
+
+    expect(parsedModule.value[contractName].init).toBeDefined();
+    expect(parsedModule.value[contractName].receive).toBeDefined();
+});
+
+test('Deserialize V2 versioned schema from file (auction)', async () => {
+    const rawModule = Buffer.from(
+        fs.readFileSync('./test/resources/auction-with-errors-schema.bin')
+    );
+    const parsedModule = deserialModuleFromBuffer(rawModule);
+
+    const contractName = 'auction';
 
     if (parsedModule.v != SchemaVersion.V2) {
         throw new Error('Unexpected schema version');
@@ -63,4 +79,7 @@ test('Deserialize V2 versioned schema from file (cis2-wccd)', async () => {
 
     expect(parsedModule.value[contractName].init).toBeDefined();
     expect(parsedModule.value[contractName].receive).toBeDefined();
+    expect(
+        parsedModule.value[contractName].receive.finalize.error
+    ).toBeDefined();
 });